Drop TARGET_IPHONE_SIMULATOR macro check
Recent versions of Clang always define these TARGET_ macros (to 0 or 1
as appropriate) for Apple targets.
https://chromium-review.googlesource.com/c/libyuv/libyuv/+/5249072
made the code correctly check the *value* of the macro rather than
whether it was defined or not.

However, that broke the code when running on the iOS simulator (where
the macro is now 1).

It seems the use of this macro was just incorrect, and the code only
worked since it was never defined at all.

The original use of the macro in this file was added in
https://chromium.googlesource.com/libyuv/libyuv/+/2c8108e6c2c6fa0c62670a70be7ef6f59cf03848
but it 's not quite clear to me why.

All other uses have subsequently been removed, e.g. in
https://chromium.googlesource.com/libyuv/libyuv/+/6a1d01220ac1d6fa33058ec2d0689fbafe0e46e1

this removes the last instance, and should fix the iOS simulator builds.

Bug: chromium:1519899
Change-Id: Iaf44d2c37086f1153096044df5d9b61797f66a4f