[AArch64] Improve ARGBTOARGB4444 using SRI instructions
The existing sequence to convert from 8-bit ARGB to 4-bit ARGB4444 makes
use of a lot of shifts and bit-clears before ORR'ing the pairs together.
This is unnecessary since we can do the same with the SRI instruction,
so use that instead.

Reduction in runtime for selected kernels:

                Kernel | Cortex-A55 | Cortex-A76
ARGBToARGB4444Row_NEON |     -15.3% |     -16.6%
I422ToARGB4444Row_NEON |      -2.7% |     -11.9%

Bug: libyuv:976
Change-Id: I86cd86c7adf1105558787a679272179821f31a9d