[AArch64] Add missing clobber, fix zero-init for compare kernels
The "memory" clobber needs to be present even if the asm does not store
anything to memory, since otherwise the compiler would be allowed to
reorder earlier stores to the pointers after they would be needed by the
asm.

Also fix up the zero-initialisation of accumulators in
SumSquareError_NEON, since EOR'ing a register by itself is not a
recognised zeroing idiom on most AArch64 micro-architectures.

Bug: libyuv:976
Change-Id: I3175367abf6f59db8371b4478f1156950277d7c5