[AArch64] Use LD1/ST1 rather than LD4/ST4 in ARGBMultiplyRow_NEON
There is no need to de-interleave channels here since we are applying
the same operation across all lanes. LD4 and ST4 are known to be
significantly slower than LD1/ST1 on some micro-architectures so we
should prefer to avoid them where possible.

Reduction in runtimes observed for ARGBMultiplyRow_NEON:

 Cortex-A55: -22.3%
Cortex-A510: -56.6%
 Cortex-A76: -45.5%
  Cortex-X2: -54.6%

Change-Id: I9103111a109a4d87d358e06eb513746314aaf66a