CMake: Improve the checks for CMAKE_SYSTEM_PROCESSOR
CMAKE_SYSTEM_PROCESSOR doesn't strictly have the values "arm" or
"aarch64", it can have more varied spellings. When cross compiling,
the value is specified by the user, but when doing native compilation,
the variable gets its value from CMAKE_HOST_SYSTEM_PROCESSOR. This
variable is defined to have the value of $(uname -m) on many Unixes.

This can give values like "armv7l" on some hosts, and "arm64" on
macOS.

Thus, when checking for aarch64, also check for the spelling "arm64".
And when checking for arm, check for any string that starts with
"arm", except for "arm64".

Change-Id: I99695ee2f3439098c897ae1408605781cd0db9fd